Исправить farm_id INTEGER -> BIGINT: реальные ID SFL превышают int32
/tickets_check падал на asyncpg.DataError ("value out of int32 range") при
чтении снэпшотов для ферм с ID вроде 1429214903236188 — farm_id в farmers
и ticket_leaderboard_snapshots был объявлен как INTEGER. Миграция 010
приводит обе колонки к BIGINT (top500_snapshots уже был BIGINT).

+++ migrations/001_farmers.sql
@@ -6,7 +6,7 @@ CREATE TABLE IF NOT EXISTS farmers (
     id                SERIAL PRIMARY KEY,
     telegram_id       BIGINT      UNIQUE NOT NULL,  -- кто привязал (необратимо)
     telegram_username TEXT,                          -- ник в ТГ (@..., может быть NULL)
-    farm_id           INTEGER     UNIQUE NOT NULL,   -- номер фермы (одна ферма = один владелец)
+    farm_id           BIGINT      UNIQUE NOT NULL,   -- номер фермы (одна ферма = один владелец)
     game_username     TEXT,                          -- ник в игре (farm.username, может быть NULL)
     xp                NUMERIC,                        -- farm.bumpkin.experience (level считает сайт)
     balance           NUMERIC,                        -- farm.balance (FLOWER)
